+++ resources/js/pages/social/components/groups/ManageApplicationsModal.vue
